cart-service/app/utils/security.py
from jose import jwt, JWTError
from app.config import settings

def decode_token(token: str) -> dict:
    try:
        return jwt.decode(token, settings.JWT_SECRET, algorithms=[settings.JWT_ALGORITHM])
    except JWTError:
        raise Exception("Invalid token")
