cart-service/requirements.txt
fastapi
uvicorn[standard]
asyncpg
alembic
sqlalchemy
pydantic
pydantic-settings
python-jose[cryptography]
python-dotenv
